+++ packages/yourloops/pages/login/index.ts
@@ -26,6 +26,6 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import Login from "./login";
+import LoginPage from "./login";
 import ConsentPage from "./consent";
-export { Login as LoginPage, ConsentPage };
+export { LoginPage, ConsentPage };