+++ packages/yourloops/pages/patient/patient-consent.tsx
@@ -26,18 +26,14 @@
  */
 
 import * as React from "react";
-import { useAuth } from "../../lib/auth";
+
 import Consent from "../../components/consent";
 
 /**
  * Patient Consent Page
  */
 function PatientConsentPage(): JSX.Element {
-  const { user } = useAuth();
-
-  return (
-    <Consent messageKey="constent-welcome-message" destinationPath={`${user?.getHomePage()}`} />
-  );
+  return <Consent messageKey="constent-welcome-message" />;
 }
 
 export default PatientConsentPage;