+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## Unrelease
+## Unreleased
 ### Fixed
+- YLP-779 Caregiver switching to hcp is logged off at the end of the process
 - YLP-Y82 Overview calendar day hover are hard to read
 
 ## 2.0.0 2021-05-21